+++ CHANGELOG.md
@@ -7,6 +7,10 @@ Added:
 - output file `niche_hidden_features.csv.gz` in `GNN_dir`, which is also available as the `niche_hidden_features` attribute in the `AnaData` class.
 - tests for IO options preparation
 
+Changed:
+
+- dir options for `ONTraC_analysis` become optional
+
 Removed:
 
 - output named `consolidate_s.csv.gz` in `GNN_dir` to avoid ambiguity. Please use `niche_level_niche_cluster.csv.gz` instead.